requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
google-genai
plotly
tqdm
wordcloud
pyperclip
